feat(thor): wire oracle tools into groups
